Use go-libp2p-gostream
The only thing that this needs to do is to provide the implementation
of roundtripper so it can be used with http, as gostream provides
the necessary Dial and Listen functionality.
